+++ lib/service/Assistant.dart
@@ -11,7 +11,7 @@ import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/service.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
-class Assistant with Service implements NotificationsListener, ContentItemCategoryClient {
+class Assistant with Service, NotificationsListener, ContentItemCategoryClient {
 
   static const String notifyFaqsContentChanged = "edu.illinois.rokwire.assistant.content.faqs.changed";
   static const String _faqContentCategory = "assistant_faqs";
@@ -20,7 +20,8 @@ class Assistant with Service implements NotificationsListener, ContentItemCatego
   Map<AssistantProvider, List<Message>> _displayMessages = <AssistantProvider, List<Message>>{
     AssistantProvider.uiuc: List<Message>.empty(growable: true),
     AssistantProvider.google: List<Message>.empty(growable: true),
-    AssistantProvider.grok: List<Message>.empty(growable: true)
+    AssistantProvider.grok: List<Message>.empty(growable: true),
+    AssistantProvider.perplexity: List<Message>.empty(growable: true)
   };
 
   // Singleton Factory
@@ -167,6 +168,8 @@ class Assistant with Service implements NotificationsListener, ContentItemCatego
     await Future.wait([
       _loadMessages(provider: AssistantProvider.uiuc),
       _loadMessages(provider: AssistantProvider.google),
+      _loadMessages(provider: AssistantProvider.grok),
+      _loadMessages(provider: AssistantProvider.perplexity),
     ]);
   }