+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Exposed ability for switching to test Illini cash url from Debug panel [#698](https://github.com/rokwire/illinois-app/issues/698).
+- Perplexity Assistant Provider [#4906](https://github.com/rokwire/illinois-app/issues/4906).
 ### Removed
 - Deleted Event2AttendanceDetailPanel (unused).
 ### Changed
 - Android: Upgrade Android Gradle Plugin to version 8.2.1.
+- Extended NotificationService APIs [#576](https://github.com/rokwire/app-flutter-plugin/issues/5726).
 ### Fixed
 - Fixed assess to Onboarding2Panel states [#4826](https://github.com/rokwire/illinois-app/issues/4826).